feature (#18): New "Conversation" system
- Dialogue system has been refactored into a new system that relies on a
directed graph of nodes
- NPCs are equipped with a Conversation script which contains the
dialog tree
- Each Conversation has an array of ConversationNode objects which
contain a speaker name, a message, a boolean isDecision, and optionally
up to three responses and nodes
- The Conversation sends the ConversationNode to the
ConversationManager, which then creates either a dialog or a decision
- If it is a decision, the ConversationManager waits for a button to be
clicked and sends the original Conversation the id of the next node
- If it is a dialog, the ConversationManager types the message and waits
for the next input of "E".
- Most files for the Dialogue system have been removed as a result of
these changes.